Consistently add meshes to a named Collection regardless of having materials or not.
As of 1.0.9, only meshes with materials are put under a Collection. Meshes that have no materials gets put under the root of the scene (edit: Scene Collection) browser. This behavior is inconsistent.

During the course of testing this: When one forgets to clear their MaterialDepot, the import will err and fail to generate a Collection. I've addressed that as well. 

See comments for lines added and changed.