Consistently add meshes to a named Collection regardless of materials. (#5)
* Consistently add meshes to a named Collection regardless of having materials or not.

As of 1.0.9, only meshes with materials are put under a Collection. Meshes that have no materials gets put under the root of the scene (edit: Scene Collection) browser. This behavior is inconsistent.

During the course of testing this: When one forgets to clear their MaterialDepot, the import will err and fail to generate a Collection. I've addressed that as well. 

See comments for lines added and changed.

* additional check, in case Builder couldn't find files in the depot

* fix indentation issue (stoopid text editor). don't want to change the basic logic

* Re-entered my changes with added stress test

scenarios tested: 
1. no material JSON
2. materiual JSON, but wrong depot path
3. Builder errors
4. (added) importing everything under base\characters\main_npc\judy

My tests indicate this no longer hangs indefinitely. 
It still takes some time similar to the main branch.