openJSON: normalize paths to support loading materials on other OSes
Even if Wolvenkit doesn't work on linux properly today, Blender does, and
getting Blender to run in a VM is hard. So it's not crazy to end up running
Wolvenkit in a VM and Blender on the Linux host.

In these situations, if you set up appropriate shared storage, you can make
all the material files available on the Linux side, but the Material json
produced by an export is using '\\' for the path separator.

This change normalises paths to use the current OS separator, so all the user
has to do is adjust the depot path in the Material json to match the other side
and materials will import. This should also help for MacOS.