skin: add support for Secondary Albedo
The secondary albedo layer is used by the VTK and UV texture frameworks as the
tattoo overlay. Including it here simplifies the process of loading an overlay
into Blender, as the user just needs to specify the image file, and doesn't
have to set up the overlay node themselves.

There are `SecondaryAlbedoInfluence` and `SecondaryAlbedoTintColorInfluence`
parameters as well, but these are not used in practice, and so I didn't bother
hooking them up, but it could be done if desired.