materialBuilder: normalise paths to work correctly on non-Windows hosts
The code that identifies material templates assumes the use of windows
paths, and so will fail to match the paths constructed when reading a
Material.json, which are normalised to the host path format.

Ideally, we'd make the material template matching code use the host
format, but it's easier to just re-normalise the paths back to windows.