multilayered: fix mask path handling on non-windows OSes
The current logic for importing layers on non-windows OSes assumes that the
various path components being used to construct the final layer path are using
Windows separators, but they are actually already using native separators.

This causes the splitting logic to fail, leading to an invalid path, and
failure to load all layers after the base one, which is always assumed to be a
fully open layer (is this is really true for all mask sets?)

So, let's simplify the logic by using `os.path.join` and `os.path.split` to
properly recombine the path elements in an OS-neutral way.